Network overfits perfectly.
Note: When trying to learn u_t and d_t to test neural stack, do not
expect the stack to learn the exact sequence of 1's and 0's, just look
at the read vector and be happy. It will overfit to some weird
combination of d_t and u_t (0-1) such that the read vector is perfect.
However, it will NOT tune u_t and d_t to be what you expect it to be.

Note2: If you only screw up 1 item in u_t and d_t while testing it, it
will give you the 1/0 correct value.

Also, use print("") instead of print "". py3 compatibility.

 Changes to be committed:
        modified:   NSTEST_learningToPush.py
        modified:   ltt/memory/neural_stack.py